Add car texture rendering and rotation

+++ src/Point.h
@@ -3,8 +3,8 @@
 
 struct Point
 {
-  int x, y;
-  Point(int x, int y) : x(x), y(y) {}
+  float x, y;
+  Point(float x, float y) : x(x), y(y) {}
 };
 
 #endif /* POINT */
